integracion pantalla y lista de productos

+++ frontend/the_life_water/lib/domain/entities/product.dart
@@ -2,7 +2,7 @@ class Product {
   final int id;
   final String nombre;
   final String description;
-  final int precio;
+  final double precio;
   final bool borrado;
 
   Product(
